各種修正
#3 ブロックを並べるスキルMineStack優先設定時MIneStackからアイテムが消費されない
#4 建築量を正確にしたい(建築量をint型ではなくBigDecimalで管理)
その他、細かな修正

+++ src/com/github/unchama/seichi/sql/PlayerDataTableManager.java
@@ -124,10 +124,10 @@ public class PlayerDataTableManager extends SeichiTableManager {
 		return ans;
 	}
 
-	public int getTotalBuildNum(GiganticPlayer gp) {
+	public double getTotalBuildNum(GiganticPlayer gp) {
 		String command = "";
 		final String struuid = gp.uuid.toString().toLowerCase();
-		int ans = 0;
+		double ans = 0;
 
 		command = "select build_count from " + db + "." + table
 				+ " where uuid = '" + struuid + "'";
